Move getRoteProps method to routes utils directory
